requirements.txt
pillow==12.0.0
PyMuPDF==1.26.5
